+++ contracts/finality/src/exec/finality.rs
@@ -137,7 +137,10 @@ fn verify_finality_signature(
     signing_context: &str,
     signature: &[u8],
 ) -> Result<(), ContractError> {
-    let proof_height = pr_commit.start_height + proof.index;
+    // For sparse generation: proof_height = start_height + index * interval
+    // For consecutive generation (interval=1): proof_height = start_height + index * 1 = start_height + index
+    let proof_height = pr_commit.start_height + proof.index * pr_commit.interval;
+
     if block_height != proof_height {
         return Err(ContractError::InvalidFinalitySigHeight(
             proof_height,
@@ -289,6 +292,7 @@ pub(crate) mod tests {
         let pr_commit = PubRandCommit {
             start_height: pr_commit.start_height,
             num_pub_rand: pr_commit.num_pub_rand,
+            interval: 1, // TODO: test with interval > 1
             babylon_epoch: current_epoch,
             commitment: pr_commit.commitment,
         };
@@ -298,9 +302,31 @@ pub(crate) mod tests {
         // This needs mock data from babylon_test_utils
         // https://github.com/babylonlabs-io/rollup-bsn-contracts/issues/66
         let context = "";
+        // For test, we need to create a mock storage with proper config
+        let mut deps = mock_deps_babylon();
+
+        // Set up a default config for testing (interval = 1 for consecutive generation)
+        let config = crate::state::config::Config {
+            bsn_id: "test-bsn".to_string(),
+            min_pub_rand: 1,
+            rate_limiting: crate::state::config::RateLimitingConfig {
+                max_msgs_per_interval: 100,
+                block_interval: 10,
+            },
+            bsn_activation_height: 0,
+            // Consecutive generation for this test
+            // TODO: test with interval > 1
+            finality_signature_interval: 1,
+        };
+        crate::state::config::set_config(deps.as_mut().storage, &config).unwrap();
+
+        // Calculate block height using the same logic as verify_finality_signature
+        let block_height = pr_commit.start_height
+            + proof.index.unsigned_abs() * config.finality_signature_interval;
+
         let res = verify_finality_signature(
             &hex::decode(&pk_hex).unwrap(),
-            pr_commit.start_height + proof.index.unsigned_abs(),
+            block_height,
             &pub_rand_value,
             // we need to add a typecast below because the provided proof is of type
             // tendermint_proto::crypto::Proof, whereas the fn expects babylon_merkle::proof