+++ CHANGELOG.md
@@ -39,10 +39,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### State and API breaking
 
-* [#82](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/108) core: 
-  add allow-list to instantiate arg
+* [#110](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/110) feat: sparse pub rand generation
 * [#97](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/97) feat:
-  versioning of FP allowlist
+* [#82](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/108) core: add allow-list to instantiate arg
 * [#100](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/100) chore:
   add query to fetch highest voted height